feat: splits core class into multiple files
